backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
openai
python-dotenv
